@@ -8,16 +8,17 @@
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.26.1",
+    "axios": "^1.8.4",
     "bootstrap": "^5.1.3",
     "emailjs-com": "^3.2.0",
-    "react": "^17.0.2",
+    "pdfjs-dist": "^5.0.375",
+    "react": "^18.3.1",
     "react-bootstrap": "^2.2.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-github-calendar": "^3.2.2",
     "react-icons": "^4.8.0",
     "react-parallax-tilt": "^1.7.42",
-    "react-pdf": "^5.7.1",
+    "react-pdf": "^9.2.1",
     "react-router-dom": "^6.2.2",
     "react-scripts": "5.0.0",
     "react-tsparticles": "^1.42.2",
@@ -51,6 +52,8 @@
     ]
   },
   "devDependencies": {
+    "eslint": "^8.57.1",
+    "eslint-plugin-react": "^7.37.4",
     "gh-pages": "^6.1.1"
   }
 }
