@@ -16,29 +16,39 @@ function Introduction() {
           </h1>
 
           {/* Paragraph 1 */}
-          <p className='bottom-body'>
-            Driven by a deep passion for technology and innovation, I am a Computer Science graduate
-            from the <b className='purple' children='University of Toronto' /> with expertise in
-            software engineering principles, full-stack development, and algorithm design. My
-            technical background spans
-            <b
-              className='purple'
-              children=' object-oriented, functional, and logical programming'
-            />
-            , allowing me to build scalable, efficient solutions for real-world challenges.
+          <p className='bottom-body' style={{ marginLeft: '5px', marginRight: '10px' }}>
+          In today’s data-driven world, businesses have access to more information than ever, yet 
+          many still struggle to translate that data into actionable solutions. The complexity
+          increases when these challenges span multiple functions, making it difficult to pinpoint
+          the root cause and determine the best course of action. That’s where I come in.
           </p>
 
           {/* Paragraph 2 */}
-          <p className='bottom-body'>
-            I’m eager to apply my software engineering and full-stack development expertise to a
-            forward-thinking team where I can continue to grow and make an impact. Let’s connect if
-            you’re looking for someone passionate about building scalable, high-quality software
-            solutions.
+          <p className='bottom-body' style={{ marginLeft: '5px', marginRight: '10px' }}>
+          With over six years of experience in finance, reporting, and analytics, I specialize in
+          navigating these complexities by transforming data into strategic insights that solve real
+          business problems. I excel at cross-functional collaboration in order to identify issues
+          and deliver data-backed solutions that align with business goals. My approach goes beyond
+          just the numbers. I am passionate about understanding customer needs and delivering
+          solutions that truly make an impact. I engage with stakeholders across departments to
+          ensure every perspective is considered, embracing diverse viewpoints and fostering an
+          inclusive environment that drives innovation. By leveraging my blend of technical
+          expertise, business acumen, and data storytelling skills, I help organizations turn
+          value across organizations.insights into decisions that deliver meaningful results. I am
+          committed to a Growth Mindset—continuously learning, adapting, and seeking innovative
+          approaches to overcome challenges. With a focus on collaboration and making a difference,
+          I help drive impactful solutions that create value across organizations.
+          </p>
+
+          {/* Paragraph 3 */}
+          <p className='bottom-body' style={{ marginLeft: '5px', marginRight: '10px' }}>
+          If your company is facing challenges that require a strategic, data-driven approach, I am
+          ready to help you find the solutions you need.
           </p>
         </Col>
 
         {/* Profile picture section */}
-        <Col md={4} className='bottom-img'>
+        <Col md={4} className='bottom-img' style={{ alignSelf: 'center'}}>
           <Tilt>
             <img src={profilePicture} className='img-fluid' alt='profilePicture' />
           </Tilt>
