@@ -4,10 +4,10 @@ import Button from 'react-bootstrap/Button'
 import { AiOutlineDownload } from 'react-icons/ai'
 import { Document, Page, pdfjs } from 'react-pdf'
 import 'react-pdf/dist/esm/Page/AnnotationLayer.css'
-import pdf from '../assets/Saad-Bhatti-Resume.pdf'
+import pdf from '../assets/Saarem-Bhatti-Resume.pdf'
 import Particle from '../components/Particle'
 import '../styles/Resume.css'
-pdfjs.GlobalWorkerOptions.workerSrc = `//cdnjs.cloudflare.com/ajax/libs/pdf.js/${pdfjs.version}/pdf.worker.min.js`
+pdfjs.GlobalWorkerOptions.workerSrc = `//cdnjs.cloudflare.com/ajax/libs/pdf.js/${pdfjs.version}/pdf.worker.min.mjs`;
 
 function Resume() {
   const [width, setWidth] = useState(1200)
@@ -32,7 +32,12 @@ function Resume() {
         {/* Resume view */}
         <Row className='resume'>
           <Document file={pdf} className='d-flex justify-content-center'>
-            <Page pageNumber={1} scale={width > 786 ? 1.7 : 0.6} />
+          <Page
+            pageNumber={1}
+            scale={width > 786 ? 1.7 : 0.6}
+            renderTextLayer={false}
+            renderAnnotationLayer={false}
+          />
           </Document>
         </Row>
 
